refactor: split s3 module service
