+++ TerraformRegistry.S3/S3ModulePurgeWorkflow.cs
@@ -1,0 +1,189 @@
+using Microsoft.Extensions.Logging;
+using TerraformRegistry.API.Interfaces;
+using TerraformRegistry.Models;
+
+namespace TerraformRegistry.S3;
+
+internal sealed class S3ModulePurgeWorkflow(
+    IDatabaseService databaseService,
+    S3ModuleObjectStore objectStore,
+    ILogger logger)
+{
+    public async Task<bool> PurgeModuleVersionAsync(string @namespace, string name, string provider, string version)
+    {
+        ModuleStorage? activeModuleStorage;
+        ModuleStorage? moduleStorage;
+
+        try
+        {
+            activeModuleStorage = await databaseService.GetModuleStorageAsync(@namespace, name, provider, version);
+            moduleStorage = activeModuleStorage ??
+                            await databaseService.GetModuleStorageIncludingDeletedAsync(@namespace, name, provider,
+                                version);
+        }
+        catch (Exception ex)
+        {
+            logger.LogError(
+                ex,
+                "Error reading module row for purge {Namespace}/{Name}/{Provider}/{Version}.",
+                @namespace,
+                name,
+                provider,
+                version);
+            return false;
+        }
+
+        if (moduleStorage == null)
+        {
+            return false;
+        }
+
+        var logicalObjectKey = S3ModuleObjectKeys.CreateLogicalObjectKey(@namespace, name, provider, version);
+        var purgeableObjectKeys = await objectStore.CollectPurgeableObjectKeysAsync(logicalObjectKey, moduleStorage);
+        if (!purgeableObjectKeys.Success)
+        {
+            return false;
+        }
+
+        return activeModuleStorage != null
+            ? await PurgeActiveModuleAsync(activeModuleStorage, moduleStorage, purgeableObjectKeys.ObjectKeys)
+            : await PurgeDeletedModuleAsync(moduleStorage, purgeableObjectKeys.ObjectKeys);
+    }
+
+    private async Task<bool> PurgeActiveModuleAsync(
+        ModuleStorage activeModuleStorage,
+        ModuleStorage moduleStorage,
+        IReadOnlyList<string> objectKeys)
+    {
+        try
+        {
+            var removed = await databaseService.RemoveModuleExactAsync(activeModuleStorage);
+            if (!removed)
+            {
+                logger.LogWarning(
+                    "Failed to remove exact active module row during purge for {Namespace}/{Name}/{Provider}/{Version}.",
+                    activeModuleStorage.Namespace,
+                    activeModuleStorage.Name,
+                    activeModuleStorage.Provider,
+                    activeModuleStorage.Version);
+                return false;
+            }
+
+            var deletedObjects = await objectStore.DeletePurgeableObjectKeysAsync(objectKeys, moduleStorage);
+            if (deletedObjects)
+            {
+                return true;
+            }
+
+            await TryRestoreActiveModuleSnapshotAsync(activeModuleStorage);
+            return false;
+        }
+        catch (Exception ex)
+        {
+            logger.LogError(
+                ex,
+                "Error removing exact active module row during purge for {Namespace}/{Name}/{Provider}/{Version}.",
+                activeModuleStorage.Namespace,
+                activeModuleStorage.Name,
+                activeModuleStorage.Provider,
+                activeModuleStorage.Version);
+            return false;
+        }
+    }
+
+    private async Task<bool> PurgeDeletedModuleAsync(ModuleStorage moduleStorage, IReadOnlyList<string> objectKeys)
+    {
+        try
+        {
+            var removed = await databaseService.RemoveDeletedModuleAsync(
+                moduleStorage.Namespace,
+                moduleStorage.Name,
+                moduleStorage.Provider,
+                moduleStorage.Version);
+            if (!removed)
+            {
+                logger.LogWarning(
+                    "Failed to remove deleted module row during purge for {Namespace}/{Name}/{Provider}/{Version}.",
+                    moduleStorage.Namespace,
+                    moduleStorage.Name,
+                    moduleStorage.Provider,
+                    moduleStorage.Version);
+                return false;
+            }
+
+            var deletedObjects = await objectStore.DeletePurgeableObjectKeysAsync(objectKeys, moduleStorage);
+            if (deletedObjects)
+            {
+                return true;
+            }
+
+            await TryRestoreDeletedModuleSnapshotAsync(moduleStorage);
+            return false;
+        }
+        catch (Exception ex)
+        {
+            logger.LogError(
+                ex,
+                "Error removing deleted module row during purge for {Namespace}/{Name}/{Provider}/{Version}.",
+                moduleStorage.Namespace,
+                moduleStorage.Name,
+                moduleStorage.Provider,
+                moduleStorage.Version);
+            return false;
+        }
+    }
+
+    private async Task TryRestoreActiveModuleSnapshotAsync(ModuleStorage module)
+    {
+        try
+        {
+            var restored = await databaseService.AddModuleAsync(module);
+            if (!restored)
+            {
+                logger.LogError(
+                    "Failed to restore active module row during purge rollback for {Namespace}/{Name}/{Provider}/{Version}.",
+                    module.Namespace,
+                    module.Name,
+                    module.Provider,
+                    module.Version);
+            }
+        }
+        catch (Exception ex)
+        {
+            logger.LogError(
+                ex,
+                "Error restoring active module row during purge rollback for {Namespace}/{Name}/{Provider}/{Version}.",
+                module.Namespace,
+                module.Name,
+                module.Provider,
+                module.Version);
+        }
+    }
+
+    private async Task TryRestoreDeletedModuleSnapshotAsync(ModuleStorage module)
+    {
+        try
+        {
+            var restored = await databaseService.AddDeletedModuleAsync(module);
+            if (!restored)
+            {
+                logger.LogError(
+                    "Failed to restore deleted module row during purge rollback for {Namespace}/{Name}/{Provider}/{Version}.",
+                    module.Namespace,
+                    module.Name,
+                    module.Provider,
+                    module.Version);
+            }
+        }
+        catch (Exception ex)
+        {
+            logger.LogError(
+                ex,
+                "Error restoring deleted module row during purge rollback for {Namespace}/{Name}/{Provider}/{Version}.",
+                module.Namespace,
+                module.Name,
+                module.Provider,
+                module.Version);
+        }
+    }
+}
